some abstraction, demo a LongConstantStorage
